requirements.txt
openai>=1.40.0
python-dotenv
termcolor  # For pretty CLI output like Cursor